+++ src/lib.rs
@@ -70,7 +70,7 @@ pub use run::{
 /// Human-readable time formatting
 pub mod time;
 pub use time::{
-	Time,TimeFull,
+	Time,TimeFull,Uptime,
 };
 
 #[cfg(feature = "date")]