+++ CHANGELOG.md
@@ -17,11 +17,15 @@ Types of changes:
 
 # readable Unreleased
 ## Added
+* `Htop` type in `readable::time` for [`htop`](https://github.com/htop-dev/htop)-style uptime formatting
+* `Uptime` trait for getting system uptime and formatting directly into a `Time`, `TimeFull` or `Htop`
 * `TryFrom` implementations for smart pointer strings (`Arc<str>`, `Box<str>`, etc)
 
 ## Changed
 * `Str::from_str()` -> `Str::from_str_exact()`
 * `Str::from_bytes()` -> `Str::from_bytes_exact()`
+* `Time::UNKNOWN` changed from `???` to `(unknown)`
+* `TimeFull::UNKNOWN` changed from `???` to `(unknown)`
 
 ---